dcmtk: Use shared libraries, add dev and man subpackages
The main goal of this commit is to reduce the size of the dcmtk package
on disk. Before this change, the "bin" folder weighs around 250MB on
disk, with shared libraries we're down to 53MB. We also introduce dev
and man subpackages to further reduce the size of teh main package.